Set sqlite3 version
I'm not sure what changed with sqlite3 but we saw this same issue on
Rails where suddenly we got errors trying to load the gem:

```
Specified 'sqlite3' for database adapter, but the gem is not loaded. Add
`gem 'sqlite3'` to your Gemfile (and ensure its version is at the
minimum required by ActiveRecord).
```

This is now fixed.