Make `controller` component smarter
This introduces a change to combine the `controller` and
`controller_with_namespace` components by making the `controller`
component rely on `controller_path` and manipulate the value that comes
back from that.

The idea for this came from an issue where we wanted the benefits of the
`controller_with_namespace` component but we already have a bunch of
logs with `controller`. Once I looked into how the controller name was
being fetched, I seen we had the chance to consolidate the two
components and remove some of the duplicated functionality.